update add domain functions

+++ lib/api/domains.ts
@@ -1,6 +1,5 @@
 import prisma from "@/lib/prisma";
-
-import { getApexDomain, removeDomainFromVercel } from "../domains";
+import { getApexDomain } from "../domains";
 
 // calculate the domainCount
 export async function getDomainCount(domain: string) {
@@ -27,21 +26,15 @@ export async function getDomainCount(domain: string) {
 export async function deleteDomain(
   domain: string,
   {
-    // Note: in certain cases, we don't need to remove the domain from the Prisma
     skipPrismaDelete = false,
   } = {},
 ) {
-  const domainCount = await getDomainCount(domain);
-
-  return await Promise.allSettled([
-    // remove the domain from Vercel
-    removeDomainFromVercel(domain, domainCount),
-    // delete domain
-    !skipPrismaDelete &&
-      prisma.domain.delete({
-        where: {
-          slug: domain,
-        },
-      }),
-  ]);
+  if (!skipPrismaDelete) {
+    await prisma.domain.delete({
+      where: {
+        slug: domain,
+      },
+    });
+  }
+  return true;
 }
